feat: moved table to a single component
