+++ src/app/app.module.ts
@@ -27,6 +27,7 @@ import { WorkersComponent } from './workers/workers.component';
 import { WorkerSearchComponent } from './workers.search/worker.search.component';
 import { ConfirmationDialogComponent } from './confirmation.dialog/confirmation.dialog.component';
 import { WorkerInputDialogComponent } from './worker-input-dialog/worker-input-dialog.component';
+import { WorkersTableComponent } from './workers.table/workers.table.component';
 
 export function HttpLoaderFactory(httpClient: HttpClient) {
     return new TranslateHttpLoader(httpClient, './assets/i18n/');
@@ -55,6 +56,7 @@ const materialsModules = [
         WorkerSearchComponent,
         ConfirmationDialogComponent,
         WorkerInputDialogComponent,
+        WorkersTableComponent,
     ],
     imports: [
         BrowserModule,
