+++ src/app/workers/workers.component.html
@@ -1,63 +1,2 @@
 <app-worker-search (worker)="searchedEmployer($event)"></app-worker-search>
-<table mat-table [dataSource]="dataSource" class="mat-elevation-z8">
-    <tr mat-header-row *matHeaderRowDef="displayedColumns; sticky: true"></tr>
-    <tr
-        mat-row
-        *matRowDef="let row; columns: displayedColumns"
-        (click)="clickedRow(row)"
-        [ngClass]="{ selectedRow: workersID.indexOf(row.id) > -1 }"
-    ></tr>
-    <ng-container matColumnDef="id">
-        <th mat-header-cell *matHeaderCellDef>id</th>
-        <td mat-cell *matCellDef="let element">
-            {{ element.id }}
-        </td>
-    </ng-container>
-    <ng-container matColumnDef="fullName">
-        <th mat-header-cell *matHeaderCellDef>
-            {{ 'USERS_TABLE.FULLNAME' | translate }}
-        </th>
-        <td mat-cell *matCellDef="let element">
-            {{ element.name + ' ' + element.surname }}
-        </td>
-    </ng-container>
-    <ng-container matColumnDef="dob">
-        <th mat-header-cell *matHeaderCellDef>
-            {{ 'USERS_TABLE.DOB' | translate }}
-        </th>
-        <td mat-cell *matCellDef="let element">
-            {{ element.dob }}
-        </td>
-    </ng-container>
-    <ng-container matColumnDef="panel">
-        <th mat-header-cell *matHeaderCellDef>
-            {{ 'USERS_TABLE.PANEL' | translate }}
-        </th>
-        <td
-            mat-cell
-            *matCellDef="let element"
-            (click)="workersID.push(element.id)"
-        >
-            <span
-                *ngIf="workersID.length >= 1 && workersID.includes(element.id)"
-            >
-                <button mat-mini-fab class="edit" (click)="openDialog(element)">
-                    <mat-icon class="mat-18">edit</mat-icon>
-                </button>
-                <button
-                    mat-mini-fab
-                    color="warn"
-                    (click)="deleteEmployer(element)"
-                >
-                    <mat-icon class="mat-18">delete</mat-icon>
-                </button>
-            </span>
-        </td>
-    </ng-container>
-</table>
-<div class="loader-container" *ngIf="isLoadingResults">
-    <mat-spinner></mat-spinner>
-</div>
-<button mat-mini-fab class="edit" (click)="openDialog()">
-    <mat-icon class="mat-18">edit</mat-icon>
-</button>
+<app-workers-table></app-workers-table>
